sim2real/requirements.txt
pyserial==3.4
pynmea2==1.19.0
click==8.1.7
scipy
